feat: implement Gemini AI financial diagnostics and add home, record creation, and reporting screens

+++ lib/services/gemini_ai_service.dart
@@ -71,9 +71,10 @@ $dataSummary
       totalIncome: totalIncome,
       totalExpense: totalExpense,
       balance: balance,
-      savingsRate: savingsRate,
+      savingsRate: savingsRate.toDouble(),
       sortedCategories: sortedCategories,
     );
+
   }
 
   static String _generateLocalHeuristicDiagnosis({
